json fixture loading, move notification tests
